storage/requirements.txt
mmh3
pynacl
